Separate out sub-package `app.types` from module `app.utils.types`
